Refactor: Remove interface validation tests and implement RBAC system
- Deleted the interface validation test file `interface_validation_test.go` which contained tests for SecretsManager and UserManager interfaces.
- Introduced a new RBAC (Role-Based Access Control) system in `rbac.go` to manage user roles and permissions.
- Added user management functionalities including user creation, deletion, and role updates.
- Implemented secure token generation and hashing for user authentication in `user.go`.
- Established methods for loading users and roles from JSON configuration files, ensuring first-run setup and validation of user data.
